Restructured by splitting script into different classes in separate files
